v3/msa-saga/requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
autopep8==1.5.7
backcall==0.2.0
bleach==3.3.0
certifi==2021.10.8
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
ConfigArgParse==1.5.1
coverage==6.3.2
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.68.0
ffmpeg==1.4
filelock==3.0.12
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
iniconfig==1.1.1
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
locust==2.0.0
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.3.0
numpy==1.20.2
packaging==21.3
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.2.0
Pillow==8.2.0
pluggy==1.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psutil==5.8.0
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pymongo==4.0.1
pyparsing==3.0.7
pyrsistent==0.17.3
pytest==7.0.1
pytest-flask==1.2.0
pytest-ordering==0.6
python-dateutil==2.8.1
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.27.1
roundrobin==0.0.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.22
starlette==0.14.2
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.2.0
tokenizers==0.10.2
toml==0.10.2
tomli==2.0.1
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
transformers==4.5.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.3
widgetsnbextension==3.5.1
zope.event==4.5.0
zope.interface==5.4.0
